Fix a Linux clang build error (#931)
Building with the clang compiler on Linux errors here,
with `sqrt()` returning a double.

+++ Source/BiquadFilter.h
@@ -68,7 +68,7 @@ public:
    void Filter(float* buffer, int bufferSize);
 
    float mF{ 4000 };
-   float mQ{ sqrt(2.0f) / 2 };
+   float mQ{ static_cast<float>(sqrt(2.0f) / 2) };
    float mDbGain{ 0 };
    FilterType mType{ FilterType::kFilterType_Lowpass };
